chore: refactor user files
