+++ client/app/(authenticated)/dashboard/profile/page.tsx
@@ -1,9 +1,9 @@
 import { Suspense } from "react";
 import Loading from "@/app/components/loading/SkeletonSpinner";
-import User from "@/app/components/routes/profile/User";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { IDP } from "@/app/utils/enums";
+import User from "@/app/components/routes/profile/Page";
 
 // 🏗️ Sync server component: dashboard\profile
 export default async function Page() {
